\name{MAMSEpo}
\alias{MAMSEmulti} 
\alias{MAMSEmultipo} 
\alias{MAMSEsurv}
\alias{MAMSEsurvpo}
\alias{MAMSEuni}
\alias{MAMSEunipo}
\title{Functions for use by MAMSE.}
\description{Functions used by \link{MAMSE} to compute the MAMSE weights.}
\keyword{nonparametric}
\keyword{survival}
\keyword{multivariate}
\keyword{univar}
