\name{ranked}
\alias{ranked}

\title{Function used by \link{MAMSE} to transform the data into rescaled ranks.}
\description{Functions used by \link{MAMSE} to convert the data into rescaled
ranks. By default, \code{ranked=function(x){ rank(x)/(length(x)+1) }}. The user
can redefine the function according to other rescalings.}
\keyword{nonparametric}
\keyword{MAMSE}
\keyword{copula}

